+++ chia/server/start_data_layer.py
@@ -18,7 +18,7 @@ from chia.util.chia_logging import initialize_logging
 from chia.util.config import load_config, load_config_cli
 from chia.util.default_root import DEFAULT_ROOT_PATH
 from chia.util.ints import uint16
-from chia.util.misc import SignalHandlers
+from chia.util.misc import SignalHandlers, load_plugin_configurations
 
 # See: https://bugs.python.org/issue29288
 "".encode("idna")
@@ -80,37 +80,6 @@ def create_data_layer_service(
         connect_to_daemon=connect_to_daemon,
     )
 
-async def load_plugin_configurations(root_path: str, config_type: str) -> List[str]:
-    """
-    Loads plugin configurations from the specified directory and validates that the contents
-    are in the expected JSON format (an array of strings). It gracefully handles errors and ensures
-    that the necessary directories exist, creating them if they do not.
-
-    Args:
-        root_path (str): The root path where the plugins directory is located.
-        config_type (str): The type of plugins to load ('downloaders' or 'uploaders').
-
-    Returns:
-        List[str]: A list of valid configurations for the specified plugin type.
-    """
-    config_path = Path(root_path) / 'plugins' / config_type
-    # Ensure the config directory exists, create if not
-    config_path.mkdir(parents=True, exist_ok=True)
-
-    valid_configs = []
-    for conf_file in config_path.glob('*.conf'):
-        try:
-            with open(conf_file, 'r') as file:
-                data = json.load(file)
-                # Validate that data is a list of strings
-                if isinstance(data, list) and all(isinstance(item, str) for item in data):
-                    valid_configs.extend(data)
-        except (IOError, json.JSONDecodeError, Exception) as e:
-            # Log or print the error based on your logging strategy
-            print(f"Error loading or parsing {conf_file}: {e}, skipping this file.")
-    return valid_configs
-
-
 async def async_main() -> int:
     # TODO: refactor to avoid the double load
     config = load_config(DEFAULT_ROOT_PATH, "config.yaml", fill_missing_services=True)
@@ -130,23 +99,24 @@ async def async_main() -> int:
     )
 
     plugins_config = config["data_layer"].get("plugins", {})
+    service_dir = DEFAULT_ROOT_PATH / SERVICE_NAME
 
     old_uploaders = config["data_layer"].get("uploaders", [])
     new_uploaders = plugins_config.get("uploaders", [])
-    file_uploaders = await load_plugin_configurations(DEFAULT_ROOT_PATH, "uploaders")
+    conf_file_uploaders = await load_plugin_configurations(service_dir, "uploaders")
     uploaders: List[PluginRemote] = [
         *(PluginRemote(url=url) for url in old_uploaders),
         *(PluginRemote.unmarshal(marshalled=marshalled) for marshalled in new_uploaders),
-        *(PluginRemote(url=url) for url in file_uploaders),
+        *(PluginRemote(url=url) for url in conf_file_uploaders),
     ]
 
     old_downloaders = config["data_layer"].get("downloaders", [])
     new_downloaders = plugins_config.get("downloaders", [])
-    file_downloaders = await load_plugin_configurations(DEFAULT_ROOT_PATH, "downloaders")
+    conf_file_downloaders = await load_plugin_configurations(service_dir, "downloaders")
     downloaders: List[PluginRemote] = [
         *(PluginRemote(url=url) for url in old_downloaders),
         *(PluginRemote.unmarshal(marshalled=marshalled) for marshalled in new_downloaders),
-        *(PluginRemote(url=url) for url in file_downloaders),
+        *(PluginRemote(url=url) for url in conf_file_downloaders),
     ]
 
     service = create_data_layer_service(DEFAULT_ROOT_PATH, config, downloaders, uploaders)
